requirements.txt
panda
pyyaml
openai
python-dotenv
tabulate
streamlit